+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change text from "is out of the stock" to "is out of stock" - @indiebytes (#3452)
 - Added general purpose hooks - @andrzejewsky (#3389)
 - Added i18n support for modules - @dz3n (#3369)
+- Added support for creating localized child routes - @gibkigonzo (#3489)
 
 ### Fixed